requirements.txt
Django==1.4.5
Markdown==2.3.1
PIL==1.1.7
South==0.8.2
boto==2.15.0
dj-database-url==0.2.2
dj-static==0.0.5
django-appconf==0.5
django-debug-toolbar==0.9.4
django-forms-bootstrap==2.0.3.post2
django-mailgun==0.2.2
django-markitup==1.0.0
django-model-utils==1.1.0
django-reversion==1.7
django-sitetree==0.9.4
django-storages==1.1.8
django-taggit==0.9.3
django-timezones==0.2
django-toolbelt==0.0.1
django-user-accounts==1.0b13
easy-thumbnails==1.2
flake8==2.0
gunicorn==17.5
html5lib==0.95
ipython==1.1.0
mccabe==0.2.1
metron==1.1
pep8==1.4.6
pinax-theme-bootstrap==3.0a9
psycopg2==2.5.1
pyflakes==0.7.3
pytz==2013b
requests==1.2.3
static==0.4
symposion==1.0b1
wsgiref==0.1.2
